Make ArbOwner only compile in the test configuration
